added whiteboard resources preparing to completely re-write solarSystem and SolarSystemManager
git-svn-id: https://svn.neumont.edu:8443/svn/cplusplus_anthony_corbin@52 c70e0d21-08ce-ef44-a38d-91c21494a941

+++ GameSolution/Game/Constroller.cpp
@@ -110,5 +110,5 @@ void Controller::draw(Core::Graphics& graphics) {
 	std::string fps = ss.str();
 	graphics.DrawString(0,0,fps.c_str());
 #endif//DEBUG_CONTROLLER
-	if(isPaused) graphics.DrawString(myWorld.getCenter().getX(),myWorld.getCenter().getY(),"GAME HAS BEEN PAUSED");
+	if(isPaused) graphics.DrawString((int)myWorld.getCenter().getX(),(int)myWorld.getCenter().getY(),"GAME HAS BEEN PAUSED");
 }
